+++ src/gui/helpers/qt_helper.py
@@ -210,21 +210,22 @@ class QtHelper:
 
 
     @staticmethod
-    def show_popup_menu(parent: QWidget, items: list[tuple[str,Callable|list]], position: QPoint, call_wrapper: Callable[[Callable,bool,bool], None]|None = None):
+    def show_popup_menu(parent: QWidget, items: list[tuple[str|tuple[str,str],Callable|list]], position: QPoint, call_wrapper: Callable[[Callable,bool,bool], None]|None = None):
         """
         Builds a menu
         Each key in `items` is the text on a menu item, or `None` to insert a separator.
         Each value in `items` is either a callable (which is called on clicking that item),
           or another dict for a submenu.
         """
-        def populate_menu(menu: QMenu, items: list[tuple[str,Callable|list]]):
-            for name,action_or_submenu in items:
-                if name is None:
+        def populate_menu(menu: QMenu, items: list[tuple[str|tuple[str,str],Callable|list]]):
+            for name_tuple,action_or_submenu in items:
+                if name_tuple is None:
                     menu.addSeparator()
                 elif isinstance(action_or_submenu, list):
-                    submenu = QtHelper.add_submenu(menu, text=name)
+                    submenu = QtHelper.add_submenu(menu, text=name_tuple)
                     populate_menu(submenu, items=action_or_submenu)
                 else:
+                    (name, shortcut) = name_tuple if isinstance(name_tuple,tuple) else (name_tuple, None)
                     bold = False
                     if name.startswith('*'):
                         name = name[1:]
@@ -239,7 +240,7 @@ class QtHelper:
                         action = make_wrapper(action_or_submenu)
                     else:
                         action = action_or_submenu
-                    QtHelper.add_menuitem(menu, text=name, action=action, bold=bold)
+                    QtHelper.add_menuitem(menu, text=name, action=action, bold=bold, shortcut=shortcut)
         menu = QMenu(parent)
         populate_menu(menu, items)
         menu.popup(position)