+++ src/info.py
@@ -2,5 +2,5 @@ class Info:
     Author = 'Andreas Pfau'
     Domain = 'apfau.de'
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.41b8"  # see https://peps.python.org/pep-0440/
-    AppDateStr = "2026-01-05"
+    AppVersionStr = "0.41b9"  # see https://peps.python.org/pep-0440/
+    AppDateStr = "2026-01-15"