+++ src/lib/expressions/sparams.py
@@ -6,6 +6,7 @@ from ..circles import StabilityCircle
 from ..utils import sanitize_filename, db2v, v2db
 from ..citi import CitiWriter
 from ..settings import Settings
+from ..file_config import FileConfig
 from .helpers import format_call_signature
 
 import skrf, math, os
@@ -450,7 +451,21 @@ class SParams:
     def plot(self, label: "str|None" = None, style: "str|None" = None, color: "str|None" = None, width: "float|None" = None, opacity: "float|None" = None):
         for sp in self.sps:
             try:
-                sp.plot(label=label, style=style, color=color, width=width, opacity=opacity, original_files=sp.original_files, param_type=sp.param_type, number_type=sp.number_type)
+                individual_color = color
+                if (individual_color is None) and (len(sp.original_files) == 1):
+                    original_file = next(iter(sp.original_files))
+                    assigned_color = FileConfig.get_color(original_file)
+                    if assigned_color:
+                        individual_color = assigned_color
+                
+                individual_style = style
+                if (individual_style is None) and (len(sp.original_files) == 1):
+                    original_file = next(iter(sp.original_files))
+                    assigned_style = FileConfig.get_style(original_file)
+                    if assigned_style:
+                        individual_style = assigned_style
+                
+                sp.plot(label=label, style=individual_style, color=individual_color, width=width, opacity=opacity, original_files=sp.original_files, param_type=sp.param_type, number_type=sp.number_type)
             except Exception as ex:
                 logging.warning(f'Plotting of <{sp.name}> failed ({ex}), ignoring')